[smscusb] Fetch MAC from device tree for Raspberry Pi Model B+
